# src/tsoa/swagger.json/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ValidationErrorResponse": {
				"properties": {
					"message": {
						"type": "string",
						"example": "Validation failed"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/FieldError"
						},
						"type": "array",
						"description": "Detailed list of field errors"
					}
				},
				"required": [
					"message",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedErrorResponse": {
				"properties": {
					"error": {
						"type": "string",
						"example": "Unauthorized error message"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForbiddenErrorResponse": {
				"properties": {
					"error": {
						"type": "string",
						"example": "Forbidden error message"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundErrorResponse": {
				"properties": {
					"error": {
						"type": "string",
						"example": "Not found Error"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TaskPriority": {
				"enum": [
					"low",
					"medium",
					"high"
				],
				"type": "string"
			},
			"TaskStatus": {
				"enum": [
					"todo",
					"progress",
					"completed"
				],
				"type": "string"
			},
			"TaskRequest": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Design the dashboard UI"
					},
					"description": {
						"type": "string",
						"example": "Create a modern and responsive UI for the admin dashboard"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority",
						"example": "low"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus",
						"example": "todo"
					},
					"startDate": {
						"type": "string",
						"example": "2025-05-10T09:00:00Z"
					},
					"endDate": {
						"type": "string",
						"example": "2025-05-15T17:00:00Z"
					},
					"estimatedHours": {
						"type": "number",
						"format": "double",
						"example": 12.5
					},
					"assignedToUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					},
					"createdByUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					},
					"projectUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					},
					"listUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					}
				},
				"required": [
					"name",
					"priority",
					"status",
					"projectUid",
					"listUid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_any_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatchTaskRequest": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Design the dashboard UI"
					},
					"description": {
						"type": "string",
						"example": "Create a modern and responsive UI for the admin dashboard"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority",
						"example": "low"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus",
						"example": "todo"
					},
					"startDate": {
						"type": "string",
						"example": "2025-05-10T09:00:00Z"
					},
					"endDate": {
						"type": "string",
						"example": "2025-05-15T17:00:00Z"
					},
					"estimatedHours": {
						"type": "number",
						"format": "double",
						"example": 12.5
					},
					"assignedToUid": {
						"type": "string",
						"example": 3
					},
					"projectUid": {
						"type": "string",
						"example": 1
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse____": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {},
						"type": "object"
					}
				},
				"required": [
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_ProjectResponse_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ProjectResponse"
					}
				},
				"required": [
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectRequest": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Website Redesign"
					},
					"description": {
						"type": "string",
						"example": "A full redesign of the company website"
					},
					"deadline": {
						"type": "string",
						"example": "2025-06-01T10:00:00.000Z"
					},
					"managerUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ProjectRequest_": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Website Redesign"
					},
					"description": {
						"type": "string",
						"example": "A full redesign of the company website"
					},
					"deadline": {
						"type": "string",
						"example": "2025-06-01T10:00:00.000Z"
					},
					"managerUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"DeleteProjectResponse": {
				"properties": {
					"message": {
						"type": "string",
						"example": "Todo deleted successfully"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.never_",
						"example": {}
					},
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					}
				},
				"required": [
					"message",
					"data",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_ListResponseData_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double",
						"example": 200
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ListResponseData"
					}
				},
				"required": [
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ListRequest": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Design the dashboard UI"
					},
					"projectUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					}
				},
				"required": [
					"name",
					"projectUid"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PatchListRequest": {
				"properties": {
					"name": {
						"type": "string",
						"example": "New list name"
					},
					"projectUid": {
						"type": "string",
						"example": "c289f43e-89de-42ef-8e1a-35a222d1ac12"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/tasks/{taskUid}": {
			"get": {
				"operationId": "GetByUid",
				"responses": {
					"200": {
						"description": "Task Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Task Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TaskRequest"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "Task Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatchTaskRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Task Deleted Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse____"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/projects": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Project Created Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Project Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_any_"
										},
										{
											"properties": {
												"total": {}
											},
											"required": [
												"total"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/projects/{projectUid}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Project Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Project Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectRequest"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "Project Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ProjectRequest_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Project Deleted Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteProjectResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/lists/{listUid}": {
			"get": {
				"operationId": "GetByUid",
				"responses": {
					"200": {
						"description": "List Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ListResponseData_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "listUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "List Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ListResponseData_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "listUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ListRequest"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "List Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ListResponseData_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "listUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatchListRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "List Deleted Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse____"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "listUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}